refactor bttv.chat-related code
