+++ src/bots.js
@@ -1,0 +1,15 @@
+module.exports = [
+    "nightbot",
+    "moobot",
+    "sourbot",
+    "xanbot",
+    "manabot",
+    "mtgbot",
+    "ackbot",
+    "baconrobot",
+    "tardisbot",
+    "deejbot",
+    "valuebot",
+    "stahpbot",
+    "mikuia"
+];
