+++ src/chat.js
@@ -1,0 +1,7 @@
+exports.tmi = require('./chat/tmi');
+exports.templates = require('./chat/templates');
+exports.takeover = require('./chat/takeover');
+exports.emotes = require('./chat/emotes');
+exports.helpers = require('./chat/helpers');
+exports.handlers = require('./chat/handlers');
+exports.store = require('./chat/store');
