+++ src/chat/takeover.js
@@ -1,0 +1,306 @@
+var vars = require('../vars'),
+    debug = require('../helpers/debug'),
+    keyCodes = require('../keycodes');
+var store = require('./store'),
+    handlers = require('./handlers'),
+    helpers = require('./helpers'),
+    rooms = require('./rooms');
+var overrideEmotes = require('../features/override-emotes'),
+    loadChatSettings = require('../features/chat-load-settings'),
+    embeddedPolling = require('../features/embedded-polling'),
+    cssLoader = require('../features/css-loader');
+
+var takeover = module.exports = function() {
+    var tmi = require('./tmi')();
+
+    if(store.isLoaded) return;
+
+    if(bttv.socketServer) {
+        if(bttv.getChannel()) helpers.lookupDisplayName(bttv.getChannel());
+        if(vars.userData.isLoggedIn) helpers.lookupDisplayName(vars.userData.login);
+    }
+
+    // Hides Group List if coming from directory
+    bttv.getChatController().set("showList", false);
+
+    if(tmi.get('isLoading')) {
+        debug.log('chat is still loading');
+        setTimeout(function() {
+            takeover();
+        }, 1000);
+        return;
+    }
+
+    // Default timestamps & mod icons to on
+    var settings = bttv.storage.getObject('chatSettings');
+    if(typeof settings.showModIcons === "undefined") {
+        settings.showModIcons = true;
+        $('.ember-chat .chat-messages').removeClass('hideModIcons');
+        bttv.storage.putObject('chatSettings', settings);
+    }
+    if(typeof settings.showTimestamps === "undefined") {
+        settings.showTimestamps = true;
+        $('.ember-chat .chat-messages').removeClass('hideTimestamps');
+        bttv.storage.putObject('chatSettings', settings);
+    }
+    if(settings.darkMode === true) {
+        settings.darkMode = false;
+        $('.chat-container').removeClass('dark');
+        bttv.storage.putObject('chatSettings', settings);
+        bttv.settings.save('darkenedMode', true);
+    }
+
+    store.isLoaded = true;
+
+    // Take over listeners
+    debug.log('Loading chat listeners');
+    for(var channel in tmi.tmiSession._rooms) {
+        if(tmi.tmiSession._rooms.hasOwnProperty(channel)) {
+            delete tmi.tmiSession._rooms[channel]._events['message'];
+            delete tmi.tmiSession._rooms[channel]._events['clearchat'];
+        }
+    }
+
+    // Handle Channel Chat
+    rooms.newRoom(bttv.getChannel());
+    tmi.tmiRoom.on('message', rooms.getRoom(bttv.getChannel()).chatHandler);
+    tmi.tmiRoom.on('clearchat', handlers.clearChat);
+    tmi.set('name', helpers.lookupDisplayName(bttv.getChannel()));
+    store.currentRoom = bttv.getChannel();
+    //tmi.tmiRoom.on('labelschanged', handlers.labelsChanged);
+
+    // Handle Group Chats
+    var privateRooms = bttv.getChatController().get('connectedPrivateGroupRooms');
+    if(privateRooms && privateRooms.length > 0) {
+        privateRooms.forEach(function(room) {
+            rooms.newRoom(room.get('id'));
+            room.tmiRoom.on('message', rooms.getRoom(room.get('id')).chatHandler);
+            room.tmiRoom.on('clearchat', handlers.clearChat);
+        });
+    }
+
+    // Load BTTV emotes if not loaded
+    overrideEmotes();
+
+    // Load Chat Settings
+    loadChatSettings();
+
+    // Hover over icons
+    $("body").off('mouseover', '.chat-line .badges .badge, .chat-line .mod-icons a').on('mouseover', '.chat-line .badges .badge, .chat-line .mod-icons a', function() {
+        $(this).tipsy({
+            trigger: 'manual',
+            gravity: "sw"
+        });
+        $(this).tipsy("show");
+    }).off('mouseout', '.chat-line .badges .badge, .chat-line .mod-icons a').on('mouseout', '.chat-line .badges .badge, .chat-line .mod-icons a', function() {
+        $(this).tipsy("hide");
+        $('div.tipsy').remove();
+    });
+
+    // hover over mod card icons
+    $("body").off('mouseover', '.bttv-mod-card button').on('mouseover', '.bttv-mod-card button', function() {
+        $(this).tipsy({
+            trigger: 'manual',
+            gravity: "s"
+        });
+        $(this).tipsy("show");
+    }).off('mouseout', '.bttv-mod-card button').on('mouseout', '.bttv-mod-card button', function() {
+        $(this).tipsy("hide");
+        $('div.tipsy').remove();
+    });
+
+    // Make Timeout/Ban/Unban buttons work and Turbo/Subscriber clickable
+    $("body").off("click", ".chat-line .mod-icons .timeout").on("click", ".chat-line .mod-icons .timeout", function() {
+        helpers.timeout($(this).parents(".chat-line").data("sender"));
+        $(this).parent().children('.ban').hide();
+        $(this).parent().children('.unban').show();
+    }).off("click", ".chat-line .mod-icons .ban").on("click", ".chat-line .mod-icons .ban", function() {
+        helpers.ban($(this).parents(".chat-line").data("sender"));
+        $(this).parent().children('.ban').hide();
+        $(this).parent().children('.unban').show();
+    }).off("click", ".chat-line .mod-icons .unban").on("click", ".chat-line .mod-icons .unban", function() {
+        helpers.unban($(this).parents(".chat-line").data("sender"));
+        $(this).parent().children('.ban').show();
+        $(this).parent().children('.unban').hide();
+    }).off("click", ".chat-line .badges .turbo, .chat-line .badges .subscriber").on("click", ".chat-line .badges .turbo, .chat-line .badges .subscriber", function() {
+        if($(this).hasClass('turbo')) {
+            window.open('/products/turbo?ref=chat_badge','_blank');
+        } else if($(this).hasClass('subscriber')) {
+            window.open(Twitch.url.subscribe(bttv.getChannel(),'in_chat_subscriber_link'),'_blank');
+        }
+    });
+
+    // Make names clickable
+    $("body").off("click", ".chat-line .from").on("click", ".chat-line .from", function() {
+        handlers.moderationCard($(this).parent().data('sender')+"", $(this));
+    });
+
+    // Give some tips to Twitch Emotes
+    if(bttv.TwitchEmoteSets && tmi.product && tmi.product.emoticons) {
+        for(var i=0; i<tmi.product.emoticons.length; i++) {
+            var emote = tmi.product.emoticons[i];
+
+            if(emote.state && emote.state === "active" && !bttv.TwitchEmoteSets[emote.emoticon_set]) {
+                bttv.socketServer.emit('give_tip', { channel: bttv.getChannel(), user: (vars.userData.isLoggedIn ? vars.userData.login : 'guest') });
+                break;
+            }
+        }
+    }
+
+    // Make chat translatable
+    if (!vars.loadedDoubleClickTranslation && bttv.settings.get("dblclickTranslation") !== false) {
+        vars.loadedDoubleClickTranslation = true;
+        $('body').on('dblclick', '.chat-line', function() {
+            helpers.translate($(this).find('.message'), $(this).data("sender"), $(this).find('.message').data("raw"));
+            $(this).find('.message').text("Translating..");
+            $('div.tipsy').remove();
+        });
+    }
+
+    var $chatInterface = $('.ember-chat .chat-interface');
+    var $chatInput = $chatInterface.find('textarea');
+    var $chatSend = $chatInterface.find('.send-chat-button');
+
+    // Disable Twitch's chat senders
+    $chatInput.off();
+    $chatSend.off();
+
+    // Message input features (tab completion, message history)
+    $chatInput.on('keyup', function(e) {
+        // '@' completion is captured only on keyup
+        if(e.which === keyCodes.Tab || e.which === keyCodes.Shift) return;
+
+        helpers.tabCompletion(e);
+    });
+
+    // Implement our own text senders (+ commands & legacy tab completion)
+    $chatInput.on('keydown', function(e) {
+        if(e.which === keyCodes.Enter) {
+            var val = $chatInput.val().trim();
+            if(e.shiftKey || !val.length) {
+                return e.preventDefault();
+            }
+
+            var $suggestions = $chatInterface.find('.suggestions');
+            if ($suggestions.length){
+                $suggestions.find('.highlighted').children().click();
+                return e.preventDefault();
+            }
+
+            if(val.charAt(0) === '/') {
+                handlers.commands(val);
+            }
+
+            // Easter Egg Kappa
+            var words = val.toLowerCase().split(' ');
+            if(words.indexOf('twitch') > -1 && words.indexOf('amazon') > -1 && words.indexOf('google') > -1) {
+                helpers.serverMessage('<img src="https://cdn.betterttv.net/special/twitchtrollsgoogle.gif"/>');
+            }
+
+            helpers.sendMessage(val);
+
+            if(bttv.settings.get('chatLineHistory') === true) {
+                if(store.chatHistory.indexOf(val) !== -1) {
+                    store.chatHistory.splice(store.chatHistory.indexOf(val), 1);
+                }
+                store.chatHistory.unshift(val);
+            }
+
+            $chatInput.val('');
+            return e.preventDefault();
+        }
+
+        var $suggestions = $chatInterface.find('.suggestions');
+        if($suggestions.length && e.which !== keyCodes.Shift) {
+            $suggestions.remove();
+        }
+
+        // Actual tabs must be captured on keydown
+        if(e.which === keyCodes.Tab) {
+            helpers.tabCompletion(e);
+            e.preventDefault();
+        }
+
+        helpers.chatLineHistory($chatInput, e);
+    });
+    $chatSend.on('click', function() {
+        var val = $chatInput.val().trim();
+        if(!val.length) return;
+
+        if(val.charAt(0) === '/') {
+            handlers.commands(val);
+        }
+
+        helpers.sendMessage(val);
+
+        if(bttv.settings.get('chatLineHistory') === true) {
+            if(store.chatHistory.indexOf(val) !== -1) {
+                store.chatHistory.splice(store.chatHistory.indexOf(val), 1);
+            }
+            store.chatHistory.unshift(val);
+        }
+
+        $chatInput.val('');
+    });
+
+    $('.ember-chat .chat-messages .chat-line').remove();
+    if(bttv.socketServer) bttv.socketServer.emit('chat history');
+    helpers.serverMessage('<center><small>BetterTTV v' + bttv.info.version + ' Loaded.</small></center>');
+    helpers.serverMessage('Welcome to '+helpers.lookupDisplayName(bttv.getChannel())+'\'s chat room!');
+
+    // Reset chatters list
+    store.chatters = {};
+    store.chatters[bttv.getChannel()] = true;
+
+    // When messages come in too fast, things get laggy
+    if(!store.__messageTimer) store.__messageTimer = setInterval(handlers.shiftQueue, 500);
+    if(!store.__lookupUsernameTimer) store.__lookupUsernameTimer = setInterval(function() { store.__usersBeingLookedUp = 0; }, 1000);
+
+    // Active Tab monitoring - Useful for knowing if a user is "watching" chat
+    $(window).off("blur focus").on("blur focus", function(e) {
+        var prevType = $(this).data("prevType");
+
+        if (prevType != e.type) {   //  reduce double fire issues
+            switch (e.type) {
+                case "blur":
+                    store.activeView = false;
+                    break;
+                case "focus":
+                    store.activeView = true;
+                    break;
+            }
+        }
+
+        $(this).data("prevType", e.type);
+    });
+
+    // Keycode to quickly timeout users
+    $(window).off("keydown").on("keydown", function(e) {
+        var keyCode = e.keyCode || e.which;
+
+        if($('.bttv-mod-card').length && bttv.settings.get("modcardsKeybinds") === true) {
+            var user = $('.bttv-mod-card').data('user');
+            switch (keyCode) {
+                case keyCodes.Esc:
+                    $('.bttv-mod-card').remove();
+                    break;
+                case keyCodes.t:
+                    helpers.timeout(user);
+                    $('.bttv-mod-card').remove();
+                    break;
+                case keyCodes.p:
+                    helpers.timeout(user, 1);
+                    $('.bttv-mod-card').remove();
+                    break;
+                case keyCodes.b:
+                    helpers.ban(user);
+                    $('.bttv-mod-card').remove();
+                    break;
+                case keyCodes.i:
+                    helpers.sendMessage("/ignore "+user);
+                    $('.bttv-mod-card').remove();
+                    break;
+            }
+        }
+    });
+}
