+++ src/chat/tmi.js
@@ -1,0 +1,3 @@
+module.exports = function() {
+	return bttv.getChatController() ? bttv.getChatController().currentRoom : false;
+}
