+++ src/features/channel-reformat/handle-resize.js
@@ -1,4 +1,4 @@
-var debug = require('../../debug'),
+var debug = require('../../helpers/debug'),
     vars = require('../../vars');
 
 var handleResize = module.exports = function () {
