+++ src/features/check-broadcast-info.js
@@ -1,4 +1,4 @@
-var debug = require('../debug');
+var debug = require('../helpers/debug');
 
 var checkBroadcastInfo = module.exports = function() {
     var channel = bttv.getChannel();
