+++ src/features/check-following.js
@@ -1,4 +1,4 @@
-var debug = require('../debug'),
+var debug = require('../helpers/debug'),
     vars = require('../vars');
 
 var checkFollowing = module.exports = function () {
