+++ src/features/highlight-feedback.js
@@ -1,4 +1,4 @@
-var debug = require('../debug');
+var debug = require('../helpers/debug');
 
 var ts_tink;
