+++ src/helpers/regex.js
@@ -1,0 +1,22 @@
+exports.escapeRegExp = function (text) {
+    // Escapes an input to make it usable for regexes
+    return text.replace(/[-[\]{}()+?.,\\^$|#\s]/g, "\\$&");
+};
+
+
+/**
+ * Function from Ryan Chatham's Twitch Chat Emotes
+ * Gets the usable emote text from a regex.
+ * @attribute http://userscripts.org/scripts/show/160183 (adaption)
+ */
+exports.getEmoteFromRegEx = function(regex) {
+    return decodeURI(regex.source)
+        .replace('&gt\\;', '>') // right angle bracket
+        .replace('&lt\\;', '<') // left angle bracket
+        .replace(/\(\?![^)]*\)/g, '') // remove negative group
+        .replace(/\(([^|])*\|?[^)]*\)/g, '$1') // pick first option from a group
+        .replace(/\[([^|])*\|?[^\]]*\]/g, '$1') // pick first character from a character group
+        .replace(/[^\\]\?/g, '') // remove optional chars
+        .replace(/^\\b|\\b$/g, '') // remove boundaries
+        .replace(/\\/g, ''); // unescape
+}
