+++ test/popup-style.styl
@@ -1,0 +1,23 @@
+.highlighter-container
+	background-color: grey
+	width: 400px
+
+	table
+		margin: 5%
+		width: 90%
+		background-color: white
+
+		thead
+			a: 1
+
+		.row1
+			background-color: lighten(grey, 40%)
+
+	.controls
+		button[disabled]
+			color: grey
+
+		//display: inline-block
+
+			.edit
+				float: right
