+++ test/popup-test.html
@@ -1,0 +1,1 @@
+<link href="popup-style.css" rel="stylesheet"/><div class="highlighter-container">The following table contains keywords which act in the <br/>same manner as the <b>Highlights</b> tab of global preferences.<br/><table><thead><tr><td>Keyword</td><td>Entry Type</td><td>Channel</td></tr></thead><tbody><tr class="row1"><td>a</td><td>idk.</td><td>idk.</td></tr><tr class="row2"><td>b</td><td>idk.</td><td>idk.</td></tr><tr class="row1"><td>c</td><td>idk.</td><td>idk.</td></tr><tr class="row2"><td>d</td><td>idk.</td><td>idk.</td></tr></tbody></table><div class="controls"><button>+</button><button disabled="disabled">-</button><button disabled="disabled" class="edit">Edit</button></div></div>
